Update main_detect.py
On line 126, there was an error in the coordinates of the transformation, as the y-axis was inverted. To fix the issue with the tf's being above the objects when they should actually be below, I simply removed the negative sign from the y-axis.
Suggestion: The axes are organized in a non-standard way. It would be helpful to standardize them to facilitate code maintenance.